[FEAT/#23] searchParam 관련 build에러 수정
